+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Develop
 Major Features and Improvements
 -------------------------------
 - add ``JohnsonSU`` PDF, the Johnson SU distribution.
+- add Landau distribution.
 
 Breaking changes
 ------------------